api/requirements.txt
anyio==3.5.0
asgiref==3.5.0
cffi==1.15.0
click==8.0.3
cryptography==36.0.1
dnspython==2.2.0
ecdsa==0.17.0
fastapi==0.73.0
h11==0.13.0
idna==3.3
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.14.1
pydantic==1.9.0
pymongo==4.0.1
python-jose==3.3.0
pytz==2021.3
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.0.1
uvicorn==0.17.4
